requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
fastapi==0.114.0
httplib2==0.22.0
idna==3.8
JPype1==1.5.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
pyparsing==3.1.4
pyreportjasper==2.1.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xmltodict==0.13.0
